webserver/requirements.txt
pyyaml
binascii
flask
WTForms
flask_wtf